requirements.txt
fastapi
uvicorn
pyyaml
pydantic
httpx
pytest